Add Google Shopping breakdown to the prompt details page
The prompt details Citations tab renders via the shared CitationsDisplay
component, which already knows how to draw the Google Shopping card — but
only when citationData.googleModule is populated. getPromptStatsFn never
built that module, and it classified citations at the domain level without
excluding Google AI Mode search/shopping surfaces, so google.com surfaces
polluted Top Cited Domains/URLs (the same bug the brand-wide view fixed in
#325).

- Extract buildGoogleModule + its types out of server/citations.ts into a
  shared lib/google-module.ts so the brand-wide and per-prompt views build
  the Google Shopping section identically.
- Rework the citation section of getPromptStatsFn to mirror the brand-wide
  view at the single-prompt level: build the googleModule, exclude Google
  surfaces from the source mix, classify at the URL level (classifyUrl +
  resolvePageType), rebuild the domain distribution from URL data, and add
  pageTypeDistribution + avgPosition.
- Drop the now-unused domain-level getPromptCitationStats query.

Closes #328
