+++ .changeset/prompt-google-shopping.md
@@ -1,0 +1,5 @@
+---
+"@workspace/web": patch
+---
+
+Added the Google Shopping breakdown to a prompt's Citations tab on the prompt details page.
