+++ apps/web/scripts/benchmark-postgres-analytics.ts
@@ -129,9 +129,6 @@ async function benchBrand(
 		results.push(await bench("getDailyCitationStats",
 			() => pgRead.getDailyCitationStats(brandId, fromDate, toDate, tz, promptIds)));
 
-		results.push(await bench("getPromptCitationStats",
-			() => pgRead.getPromptCitationStats(firstPromptId, fromDate, toDate, tz)));
-
 		results.push(await bench("getPromptCitationUrlStats",
 			() => pgRead.getPromptCitationUrlStats(firstPromptId, fromDate, toDate, tz)));
 	}
