+++ apps/web/src/server/prompts.ts
@@ -11,7 +11,6 @@ import { eq, and, desc, gte, count, sql } from "drizzle-orm";
 import {
 	getPromptsSummary,
 	getPromptsFirstEvaluatedAt,
-	getPromptCitationStats,
 	getPromptCitationUrlStats,
 	getPromptDailyStats,
 	getPromptCompetitorDailyStats,
@@ -22,8 +21,17 @@ import { generateDateRange } from "@/lib/chart-utils";
 import type { LookbackPeriod } from "@/lib/chart-utils";
 import { getEffectiveBrandedStatus, computeSystemTags } from "@workspace/lib/tag-utils";
 import { createMultiplePromptJobSchedulers } from "@/lib/job-scheduler";
-import { extractDomain, normalizeUrl, emptyCategoryCounts } from "@/lib/domain-categories";
-import { categorizeDomain } from "@/lib/domain-categories.server";
+import {
+	extractDomain,
+	normalizeUrl,
+	emptyCategoryCounts,
+	emptyPageTypeCounts,
+	resolvePageType,
+	isGoogleSurfaceUrl,
+	CITATION_PAGE_TYPES,
+} from "@/lib/domain-categories";
+import { classifyUrl } from "@/lib/domain-categories.server";
+import { buildGoogleModule } from "@/lib/google-module";
 // Server Functions
 // ============================================================================
 
@@ -401,10 +409,14 @@ export const getPromptStatsFn = createServerFn({ method: "GET" })
 		};
 
 		// ---- Citation stats ----
+		// Mirrors the brand-wide citations view (server/citations.ts) at the single-
+		// prompt level: classify each citation at the URL level, pull Google AI Mode
+		// search/shopping surfaces OUT of the source mix into a dedicated Google
+		// Shopping module, and rebuild the domain distribution from the URL data.
 		let citationStats = undefined;
 		const [brandInfo, competitorsList] = await Promise.all([
-			db.select({ website: brands.website, additionalDomains: brands.additionalDomains }).from(brands).where(eq(brands.id, prompt[0].brandId)).limit(1),
-			db.select({ domains: competitors.domains }).from(competitors).where(eq(competitors.brandId, prompt[0].brandId)),
+			db.select({ name: brands.name, website: brands.website, additionalDomains: brands.additionalDomains }).from(brands).where(eq(brands.id, prompt[0].brandId)).limit(1),
+			db.select({ id: competitors.id, name: competitors.name, domains: competitors.domains }).from(competitors).where(eq(competitors.brandId, prompt[0].brandId)),
 		]);
 
 		const primaryBrandDomain = brandInfo[0] ? extractDomain(brandInfo[0].website) : "";
@@ -412,40 +424,78 @@ export const getPromptStatsFn = createServerFn({ method: "GET" })
 		const brandDomains = new Set([primaryBrandDomain, ...additionalBrandDomains].filter(Boolean));
 		const competitorDomains = new Set(competitorsList.flatMap((c) => c.domains.map(extractDomain)).filter(Boolean));
 
-		const [domainStats, urlStats] = await Promise.all([
-			getPromptCitationStats(data.promptId, fromDateStr, toDateStr, timezone),
-			getPromptCitationUrlStats(data.promptId, fromDateStr, toDateStr, timezone),
-		]);
+		const urlStats = await getPromptCitationUrlStats(data.promptId, fromDateStr, toDateStr, timezone);
 
-		if (domainStats.length > 0) {
-			const domainDistribution = domainStats.map(({ domain, count: cnt }) => ({
-				domain,
-				count: Number(cnt),
-				category: categorizeDomain(domain, brandDomains, competitorDomains),
-			}));
+		if (urlStats.length > 0) {
+			// Google AI Mode module: Shopping products (brand vs competitor) + search
+			// queries. Built from the raw URL rows (it picks out the Google surfaces);
+			// those same surfaces are excluded from the source mix below.
+			const googleModule = buildGoogleModule(
+				urlStats.map((u) => ({ prompt_id: data.promptId, url: u.url, domain: u.domain, title: u.title, count: u.count })),
+				brandInfo[0]?.name ?? "",
+				competitorsList.map((c) => ({ id: c.id, name: c.name })),
+				() => prompt[0].value,
+			);
 
-			const urlCounts = new Map<string, { count: number; title?: string; domain: string }>();
-			for (const { url, domain, title, count: cnt } of urlStats) {
+			const urlCounts = new Map<string, { count: number; title?: string; domain: string; positionSum: number; positionCount: number }>();
+			for (const { url, domain, title, count: cnt, avg_position } of urlStats) {
+				if (isGoogleSurfaceUrl(url)) continue;
 				const normalized = normalizeUrl(url);
+				const c = Number(cnt);
+				const positionSum = avg_position != null ? Number(avg_position) * c : 0;
+				const positionCount = avg_position != null ? c : 0;
 				const existing = urlCounts.get(normalized);
 				if (existing) {
-					existing.count += Number(cnt);
+					existing.count += c;
+					existing.positionSum += positionSum;
+					existing.positionCount += positionCount;
 					if (!existing.title && title) existing.title = title;
 				} else {
-					urlCounts.set(normalized, { count: Number(cnt), title: title || undefined, domain });
+					urlCounts.set(normalized, { count: c, title: title || undefined, domain, positionSum, positionCount });
 				}
 			}
 
 			const specificUrls = Array.from(urlCounts.entries())
-				.map(([url, { count: cnt, title, domain }]) => ({
-					url, title, domain, count: cnt,
-					category: categorizeDomain(domain, brandDomains, competitorDomains),
-				}))
+				.map(([url, { count: cnt, title, domain, positionSum, positionCount }]) => {
+					const category = classifyUrl(domain, url, title, brandDomains, competitorDomains);
+					return {
+						url, title, domain, count: cnt, category,
+						pageType: resolvePageType(url, title, category),
+						avgPosition: positionCount > 0 ? Math.round((positionSum / positionCount) * 10) / 10 : null,
+					};
+				})
+				.sort((a, b) => b.count - a.count);
+
+			// Domain distribution rebuilt from URL-level data, each domain taking its
+			// category from its top-cited URL (matches the brand-wide view).
+			const domainAgg = new Map<string, { count: number; category: (typeof specificUrls)[number]["category"]; topCount: number; exampleTitle?: string }>();
+			for (const u of specificUrls) {
+				const cur = domainAgg.get(u.domain);
+				if (cur) {
+					cur.count += u.count;
+					if (u.count > cur.topCount) {
+						cur.topCount = u.count;
+						cur.category = u.category;
+						cur.exampleTitle = u.title;
+					}
+				} else {
+					domainAgg.set(u.domain, { count: u.count, category: u.category, topCount: u.count, exampleTitle: u.title });
+				}
+			}
+			const domainDistribution = Array.from(domainAgg.entries())
+				.map(([domain, v]) => ({ domain, count: v.count, category: v.category, exampleTitle: v.exampleTitle }))
 				.sort((a, b) => b.count - a.count);
 
 			const categoryCounts = emptyCategoryCounts();
-			for (const d of domainDistribution) categoryCounts[d.category] += d.count;
+			const pageTypeCounts = emptyPageTypeCounts();
+			for (const u of specificUrls) {
+				categoryCounts[u.category] += u.count;
+				pageTypeCounts[u.pageType] += u.count;
+			}
 			const totalCitations = domainDistribution.reduce((s, d) => s + d.count, 0);
+			const pageTypeDistribution = CITATION_PAGE_TYPES
+				.map((pageType) => ({ pageType, count: pageTypeCounts[pageType] }))
+				.filter((d) => d.count > 0);
 
 			if (totalCitations > 0) {
 				citationStats = {
@@ -454,6 +504,8 @@ export const getPromptStatsFn = createServerFn({ method: "GET" })
 					categoryCounts,
 					domainDistribution,
 					specificUrls,
+					pageTypeDistribution,
+					googleModule,
 				};
 			}
 		}
